// generated/src/app/api-models/submitter-profile-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SubmitterProfileMaster {
  submitterId: string;
  submitterName: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  contactName: string;
  contactTitle: string;
  phoneNumber: string;
  faxNumber: string;
  externalReference: string;
  computerDialupNo: string;
  modemMake: string;
  modemModel: string;
  modemSpeed: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  userDefined6: string;
  userDefined7: string;
  country: string;
}